git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/javascool@1231 79b98d2d-744f-0410-929f-8ac7b80d50f7

+++ javascool_V3/trunk/org.javascool/lib/run.sh
@@ -4,7 +4,7 @@
 export PATH=/usr/java/jdk1.6.0_22/bin:$PATH
 
 # Usage of this script
-if [ -z "$1" -o "$1" = "-h" -o "$1" = "--help" ] 
+if [ "$1" = "-h" -o "$1" = "--help" ] 
 then cat <<EOD
 Usage: $0 [-rejar] <main> [arguments]
  Runs one of the main (ex: org.javascool.JsMain)
@@ -20,15 +20,15 @@ fi
 
 # Retrieve javascool root directory and main class
 if echo $0 | grep '^/' ; then root="$0" ; else root="`pwd`/$0" ; fi ; root=`echo $root | sed 's/\/lib\/run.sh$//' | sed 's/\/lib\/\.\/run.sh$//'`
+JAR=`grep 'JAR *= *' $root/makefile | sed 's/.*= *\([^ ]*\) */\1/'`
 
 # Running the saxon translator
 if [ "$1" = sax ] ; then shift ; java -jar $root/lib/saxon.jar $* ; exit $? ; fi
 
 # Recompile if required
-if [ "$1" = -rejar ] ; then shift ; make -C $root jar ; if [ $? != 0 ] ; then exit $? ; fi ; fi
+if [ "$1" = -rejar -o \! -f $root/www/javascool$JAR.jar ] ; then shift ; make -C $root jar ; if [ $? != 0 ] ; then exit $? ; fi ; fi
 
 # Running the class fromn the main jar
-JAR=`grep 'JAR *= *' $root/makefile | sed 's/.*= *\([^ ]*\) */\1/'`
 if [ -z "$1" ] 
 then java -jar $root/www/javascool$JAR.jar
 else java -cp $root/www/javascool$JAR.jar $*
